+++ unicode_to_png/unicode_utils.py
@@ -1,3 +1,12 @@
+#
+# This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
+# If a copy of the MPL was not distributed with this file, You can obtain one at https://mozilla.org/MPL/2.0/
+#
+# Original Author: Sergio Palma Hidalgo
+# Project URL: https://github.com/del-Pacifico/unicode-to-png
+# Copyright (c) 2025 Sergio Palma Hidalgo
+# All rights reserved.
+#
 """Unicode classification and layout helpers for emoji rendering."""
 
 from .logging_utils import log